% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_score.R
\name{bootstrapScore}
\alias{bootstrapScore}
\title{Title}
\usage{
bootstrapScore(
  geneNetwork,
  rankFun,
  scoreFun,
  expression,
  phenotype1,
  phenotype2,
  bootstrapIterations = 1000,
  replace = TRUE
)
}
\arguments{
\item{seed}{}
}
\description{
Title
}
