% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_score.R
\name{compare_phenotypes_single}
\alias{compare_phenotypes_single}
\title{Compare two phenotypes for a single gene network}
\usage{
compare_phenotypes_single(rank_matrix, phenotype1, phenotype2, score_fun)
}
\arguments{
\item{rank_matrix}{Numeric matrix With samples as the columns. See details
for further information.}

\item{phenotype1, phenotype2}{Index vector represent the indices of each
phenotype within the \code{rank_matrix}.}

\item{score_fun}{Function which calculates a score form the
rank matrix.}
}
\value{
\code{numeric} Absolute difference of the scores for the two phenotypes
}
\description{
Compare two phenotypes for a single gene network
}
\details{
Rank matrix is normally generated from a passed rank_fun function, so the
columns should represent samples, but what the rows represent depends
on the specific rank function. For RACE, CRANE, and INFER the rows simply
represent genes, while for DIRAC they represent specific pairwise
comparison.

The phenotype vectors can contain any values that can be used with the
\code{[} matrix function, and the \code{c} function.
}
