% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_score.R
\name{comparePhenotypes}
\alias{comparePhenotypes}
\title{Title}
\usage{
comparePhenotypes(
  expression,
  geneNetworkList,
  phenoytpe1,
  phenotype2,
  rankFun,
  scoreFun,
  bootstrapIterations = 1000,
  replace = TRUE,
  asFrame = TRUE
)
}
\arguments{
\item{as_frame}{}
}
\description{
Title
}
