% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareRankEntropy.R
\name{compareRankEntropy}
\alias{compareRankEntropy}
\title{Compare Phenotypes using Rank Entropy Methods}
\usage{
compareRankEntropy(
  expression,
  method = c("DIRAC", "RACE", "CRANE", "INFER"),
  phenotype1,
  phenotype2,
  geneNetworkList,
  assayName = "counts",
  bootstrapIterations = 1000,
  replace = TRUE,
  asFrame = TRUE,
  BPPARAM = bpparam()
)
}
\arguments{
\item{expression}{Gene expression data, can be either a SummarizedExperiment
object, or an array type object which implements [], and also
rownames/colnames if phenotypes or geneNetworks use character vector
indices}

\item{method}{Method to use to calculate the rank entropy, can be DIRAC,
RACE, CRANE, or INFER}

\item{phenotype1, phenotype2}{Index vectors for the locations within
expression of the samples within each phenotype, can be a numeric,
character, or logical vector}

\item{geneNetworkList}{(Named) list of index vectors, each entry must be
a character, numeric, or logical vector describing the locations
within expression of the genes within the network}

\item{assayName}{Name of assay within SummarizedExperiment object holding
gene expression data}

\item{bootstrapIterations}{Number of iterations to perform for boostrapping
the null distribution for computing the p-value}

\item{replace}{Whether sampling during bootstrapping should be done with
replacement}

\item{asFrame}{Whether the return value should be a data.frame, otherwise
a named list is returned}

\item{BPPARAM}{Biocparallel parameter for choosing a parallel backend
to use for bootstrapping computations}
}
\value{
Either a data.frame, or a named list depending on the asFrame
parameter. In the data.frame form, with columns for geneNetwork, p1Score,
p2Score, absoluteDifference, and pValue. The list return is named
according to the names in the geneNetworkList parameter, with each value
being a named list with p1Score, p2Score, absoluteDifference, and pValue.
}
\description{
Compare the rank entropy between two phenotypes for a given gene network,
using either DIRAC, CRANE, RACE, or INFER
}
