% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crane.R
\name{crane.compare_phenotypes}
\alias{crane.compare_phenotypes}
\title{Compare phenotypes using a list of Gene Networks}
\usage{
crane.compare_phenotypes(
  gene_network,
  expression,
  phenotype1,
  phenotype2,
  bootstrap_iterations = 1000,
  parallel = TRUE,
  cores = 4,
  replace = TRUE,
  seed = NULL
)
}
\arguments{
\item{gene_network}{Integer vector containing the indices of the genes
within the network}

\item{expression}{Numeric matrix representing gene expression, rows
represent genes, and columns represent samples}

\item{phenotype1, phenotype2}{Integer vectors of indices of the samples
in each phenotype.}

\item{bootstrap_iterations}{Integer number of iterations to perform
for the bootstrapping}

\item{parallel}{Boolean whether the calculation should be done in
parallel}

\item{cores}{Integer number of cores to use for parallel computation}

\item{replace}{Boolean whether to use replacement when sampling during
bootstrap operations.}

\item{seed}{Integer seed for RNG}
}
\value{
Named list, with p1.mean_centroid_distance,
p2.mean_centroid_distance, absolute_difference, and p.value
}
\description{
Compare phenotypes using a list of Gene Networks
}
