% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crane.R
\name{crane.compare_phenotypes.shuffle}
\alias{crane.compare_phenotypes.shuffle}
\title{Compare shuffled phenotypes using mean rank centroid distance}
\usage{
crane.compare_phenotypes.shuffle(
  i,
  rank_matrix,
  combined,
  p1.size,
  p2.size,
  replace = TRUE
)
}
\arguments{
\item{i}{Integer, bootstrap iteration}

\item{rank_matrix}{Integer matrix, ranks of gene expression with genes as
rows, and samples as columns}

\item{combined}{Integer vector, indices of both phenotypes}

\item{p1.size, p2.size}{Integers, number of samples in each phenotype}

\item{replace}{Boolean whether replacement should be used when sampling}
}
\value{
Double, the absolute difference between the mean centroid
distance for the shuffled phenotypes
}
\description{
Compare shuffled phenotypes using mean rank centroid distance
}
