% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crane.R
\name{crane.compare_phenotypes.single}
\alias{crane.compare_phenotypes.single}
\title{Compare phenotypes using the mean rank centroid distance}
\usage{
crane.compare_phenotypes.single(rank_matrix, phenotype1, phenotype2)
}
\arguments{
\item{rank_matrix}{Integer matrix with genes as rows, and samples as columns}

\item{phenotype1, phenotype2}{Integer vector of indices for samples in a
phenotype}
}
\value{
Double, absolute difference between the mean centroid distance
for the two phenotypes
}
\description{
Compare phenotypes using the mean rank centroid distance
}
