% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crane.R
\name{crane.mean_centroid_distance}
\alias{crane.mean_centroid_distance}
\title{Compute the mean centroid distance}
\usage{
crane.mean_centroid_distance(rank_matrix)
}
\arguments{
\item{rank_matrix}{Integer matrix with genes as rows, and samples as columns}
}
\value{
Double, mean distance of each samples rank vector from the centroid
of all the samples rank vectors
}
\description{
Compute the mean centroid distance
}
