% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crane.R
\name{crane.rank_matrix}
\alias{crane.rank_matrix}
\title{Compute the rank matrix}
\usage{
crane.rank_matrix(expression.filtered)
}
\arguments{
\item{expression.filtered}{Numeric matrix of gene expression values, genes
are rows, and samples are columns}
}
\value{
Integer matrix of ranks
}
\description{
Compute the rank matrix
}
