% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crane.R
\name{CRANE}
\alias{CRANE}
\title{Compare phenotypes using a list of Gene Networks}
\usage{
CRANE(
  expression,
  gene_network_list,
  phenotype1,
  phenotype2,
  bootstrap_iterations = 1000,
  parallel = TRUE,
  cores = 4,
  replace = TRUE,
  seed = NULL,
  as.frame = TRUE
)
}
\arguments{
\item{expression}{Numeric matrix representing gene expression, rows
represent genes, and columns represent samples}

\item{gene_network_list}{Named list, with each element being a integer
vector containing the indices of the genes within the network}

\item{phenotype1, phenotype2}{Integer vectors of indices of the samples
in each phenotype.}

\item{bootstrap_iterations}{Integer number of iterations to perform
for the bootstrapping}

\item{parallel}{Boolean whether the calculation should be done in
parallel}

\item{cores}{Integer number of cores to use for parallel computation}

\item{replace}{Boolean whether to use replacement when sampling during
bootstrap operations.}

\item{seed}{Integer seed for RNG}

\item{as.frame}{Boolean whether return should be data.frame or named
list}
}
\value{
data.frame or named list
If as.frame is TRUE:
Returns a dataframe where each row is a gene network, and the columns
are gene_network, p1.mean_centroid_distance,
p2.mean_centroid_distance, absolute_difference, and p.value.
If as.frame is FALSE:
Returns a named list, where the names are the gene_network,
taken from the gene_network_list names. Each element is another
named list, with p1.mean_centroid_distance,
p2.mean_centroid_distance, absolute_difference, and p.value.
}
\description{
Compare phenotypes using a list of Gene Networks
}
