% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crane.R
\name{craneSampleScore}
\alias{craneSampleScore}
\title{Calculate the by sample entropy scores using the CRANE method}
\usage{
craneSampleScore(filteredExpression)
}
\arguments{
\item{filteredExpression}{Numeric Matrix, rows representing genes in
the gene network, columns representing samples in the phenotype}
}
\value{
Numeric vector, with a score for the entropy of each sample
}
\description{
Calculate the by sample entropy scores using the CRANE method
}
