% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crane.R
\name{craneScoreFunction}
\alias{craneScoreFunction}
\title{Function to score a rank matrix using the CRANE method, the average distance
between samples and the rank centroid}
\usage{
craneScoreFunction(rankMatrix)
}
\arguments{
\item{rankMatrix}{Numeric matrix, a matrix where each value represents
the rank of the gene's expression within the gene network}
}
\value{
Numeric score, value representing the mean distance to the rank
centroid
}
\description{
Function to score a rank matrix using the CRANE method, the average distance
between samples and the rank centroid
}
