% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{createClassifier}
\alias{createClassifier}
\title{Create a Classifier using one of the Rank Entropy Algorithms}
\usage{
createClassifier(
  expression,
  phenotype1,
  phenotype2,
  geneNetwork,
  method = c("DIRAC", "CRANE", "RACE"),
  phenotypeNames = c("p1", "p2"),
  geneNetworkName = "gn",
  assayName = "counts"
)
}
\arguments{
\item{expression}{Gene expression data, can be either a SummarizedExperiment
object, or an array type object which implements [], and also
rownames/colnames if phenotypes or geneNetworks use character vector
indices}

\item{phenotype1, phenotype2}{Index vectors for the locations within
expression of the samples within each phenotype, can be a numeric,
character, or logical vector}

\item{geneNetwork}{Index vector for the gene network of interest,
can be a logical, numeric or character vector}

\item{method}{Method to use to calculate the rank entropy, can be DIRAC,
RACE, CRANE, or INFER}

\item{geneNetworkName}{Name of the gene network, used for
creating default name for new colData column in the SummarizedExperiment
input (if a SummarizedExperiment is used as input)}

\item{assayName}{Name of the assay containing gene expression data in the
SummarizedExperiment object (if a SummarizedExperiment is used as input)}

\item{phenotype1Name, phenotype2Name}{Names of the phenotypes, used for
creating default name for new colData column in the SummarizedExperiment
input (if a SummarizedExperiment is used as input)}
}
\value{
A classifier function
If the input is not a SummarizedExperiment object:
The classification function will take a gene expression array,
with rows representing different genes, and columns representing
different samples. The function will return a logical vector
the same length as the number of columns in the input array,
with TRUE meaning classified as phenotype1, and FALSE meaning
classified as phenotype2
If the input is a SummarizedExperiment object:
The function will take a SummarizedExperiment as input,
along with optional parameters for assayName and ColDataName.
The assayName argument is used to determine which assay to use
as input for the classifier (defaults to the same as assayName
argument to this function). The colDataName is used to name the
output column in the returned SummarizedExperiment object.
The return will be a SummarizedExperiment object with all the same
data as the input, except with an added column in the colData
DataFrame, named according to colDataName (or
geneNetworkName_phenotype1Name if no argument is provided) which is
logical, with TRUE representing that sample being classified as
phenotype1, and FALSE representing that sample being classified as
phenotype2
}
\description{
Create a Classifier using one of the Rank Entropy Algorithms
}
