% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen.R
\name{datagen.ordered.vector}
\alias{datagen.ordered.vector}
\title{Generate ordered expression vector from a provided distribution function}
\usage{
datagen.ordered.vector(ngenes, dist, ...)
}
\arguments{
\item{ngenes}{Integer, number of genes to generate expression values for.}

\item{dist}{Function, distribution function (rnorm, rnbinom, etc.), which
will be passed the ngenes, and other keyword arguments, and should
return a numeric vector of length ngenes.}

\item{...}{Keyword arguments passed to distribution function}
}
\value{
Numeric vector of ordered gene expression values
}
\description{
Generate ordered expression vector from a provided distribution function
}
