% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen.R
\name{datagen.unordered.matrix}
\alias{datagen.unordered.matrix}
\title{Generate unordered expression matrix from a provided distribution function}
\usage{
datagen.unordered.matrix(ngenes, nsamples, dist, ...)
}
\arguments{
\item{ngenes}{Integer, number of genes to generate expression values for.}

\item{nsamples}{Integer, number of samples to generate expression values for.}

\item{dist}{Function, distribution function (rnorm, rnbinom, etc.), which
will be passed the ngenes, and other keyword arguments, and should
return a numeric vector of length ngenes.}

\item{...}{Keyword arguments passed to distribution function}
}
\value{
Numeric matrix of unordered gene expression values, with rows
representing genes, and columns samples
}
\description{
Generate unordered expression matrix from a provided distribution function
}
