% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen.R
\name{datagenDropout}
\alias{datagenDropout}
\title{Dropout a proportion of expression data}
\usage{
datagenDropout(expression, droupoutRate)
}
\arguments{
\item{expression}{Numeric matrix, represents gene expression data with genes
as rows, and samples as columns.}

\item{dropoutRate}{Double, representing the proportion of values to set to
0, 0<=dropout.rate<=1.}
}
\value{
Numeric matrix, expression data with dropout.rate proportion of
values set to 0.
}
\description{
Sets dropout.rate proportion of data to 0.
}
