% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen.R
\name{datagenGenerate}
\alias{datagenGenerate}
\title{Generates gene expression data with ordered sub network}
\usage{
datagenGenerate(
  ngenesOrdered,
  ngenesTotal,
  nsamplesOrdered,
  nsamplesTotal,
  dist,
  reorderGenes = TRUE,
  ...
)
}
\arguments{
\item{ngenesOrdered}{Integer, number of genes in the ordered sub network.}

\item{ngenesTotal}{Integer, total number of genes desired.}

\item{nsamplesOrdered}{Integer, number of samples in the ordered phenotype.}

\item{nsamplesTotal}{Integer, total number of samples}

\item{dist}{Function, distribution to use for generating data. It is passed
ngenes (for number of values to generate), and any other keyword
arguments. Must return a numeric vector.}

\item{reorderGenes}{Boolean, whether the order of genes within the ordered
sub network should be shuffled. If TRUE, the order of the genes in the
network will be randomly reshuffled. If FALSE, the genes within the network
will always have their expression values increase with increasing row
index.}

\item{...}{Keyword arguments, passed to the dist function}
}
\value{
Named list with:
expression: Numeric matrix of gene expression values with an ordered
sub network. There will be ngenes.total rows representing genes, and
nsamples.total columns representing samples.
orderedGenes: Integer vector with the indices of the genes within
the ordered sub network. These will be in ascending order, so the gene
with the lowest expression values will be in the row represented by the
first index, and the gene with the highest expression will be
represented by the last index.
orderedSamples: Integer vector containing the indices of the samples
within the ordered phenotype.
}
\description{
Generates gene expression data with ordered sub network
}
