% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen.R
\name{datagenNoise}
\alias{datagenNoise}
\title{Adds noise from a distribution to the expression data}
\usage{
datagenNoise(expression, dist, ...)
}
\arguments{
\item{expression}{Numeric matrix, gene expression data. Rows are genes,
columns are samples.}

\item{dist}{Function, distribution function to generate random numbers to be
added as noise to expression. Must take number of random numbers to
generate as first argument, and is also passed other keyword arguments.}

\item{...}{Keyword arguments, passed to dist function.}
}
\value{
Numeric matrix, expression data with noise added.
}
\description{
Adds noise from a distribution to the expression data
}
