% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen.R
\name{datagenOrderedMatrix}
\alias{datagenOrderedMatrix}
\title{Generate ordered expression matrix from a provided distribution function}
\usage{
datagenOrderedMatrix(ngenes, nsamples, dist, ...)
}
\arguments{
\item{ngenes}{Integer, number of genes to generate expression values for.}

\item{nsamples}{Integer, number of samples to generate expression values for.}

\item{dist}{Function, distribution function (rnorm, rnbinom, etc.), which
will be passed the ngenes, and other keyword arguments, and should
return a numeric vector of length ngenes.}

\item{...}{Keyword arguments passed to distribution function}
}
\value{
Numeric matrix of ordered gene expression values, with rows
representing genes, and columns samples
}
\description{
Generate ordered expression matrix from a provided distribution function
}
