% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen.R
\name{datagenUnorderedVector}
\alias{datagenUnorderedVector}
\title{Generate unordered expression vector from a provided distribution function}
\usage{
datagenUnorderedVector(ngenes, dist, ...)
}
\arguments{
\item{ngenes}{Integer, number of genes to generate expression values for.}

\item{dist}{Function, distribution function (rnorm, rnbinom, etc.), which
will be passed the ngenes, and other keyword arguments, and should
return a numeric vector of length ngenes.}

\item{...}{Keyword arguments passed to distribution function}
}
\value{
Numeric vector of unordered gene expression values
}
\description{
Generate unordered expression vector from a provided distribution function
}
