% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirac.R
\name{DIRAC.classifier}
\alias{DIRAC.classifier}
\title{Creates a phenotype classifier based on DIRAC}
\usage{
DIRAC.classifier(
  expression,
  phenotype1,
  phenotype2,
  gene_index,
  return = "is phenotype1",
  names = c("phenotype1", "phenotype2")
)
}
\arguments{
\item{expression}{Numeric matrix of gene expression values, rows are genes
columns are samples}

\item{phenotype1, phenotype2}{Numeric vectors with indices for each of the
phenotypes}

\item{gene_index}{Numeric vector of indices for the genes within the gene
network being used for classification}

\item{return}{String describing desired return value from created function:
"name":  returns a character vector of the names of the predicted
phenotypes
"is phenotype1": returns a boolean vector indicating if the
predicted phenotype is phenotype1
"is phenotype2": returns a boolean vector indicating if the predicted
phenotype is phenotype2
"phenotype number": returns numeric vector, with 1 representing the
predicted phenotype is phenotype 1, and 2 representing that the
predicted phenotype is phenotype2
Any other value will lead the function to just return the rank matching
difference}

\item{names}{optional Character vector with the desired names for the phenotypes,}
}
\value{
A function which takes as an argument a gene expression numeric
matrix with genes as the rows and samples as the columns, and returns
a vector predicting which samples are from which phenotype (see return
argument).
}
\description{
\code{DIRAC.classifier} Creates a function which takes in a matrix of gene
expression values and returns a classification of phenotype1, or phenotype2
for each of the samples.
}
\examples{
# example code
}
