% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirac.R
\name{dirac.compare_phenotype}
\alias{dirac.compare_phenotype}
\title{Compare two phenotypes using the difference in rank conservation indices}
\usage{
dirac.compare_phenotype(
  gene_index,
  expression,
  phenotype1,
  phenotype2,
  bootstrap_iterations = 1000,
  parallel = TRUE,
  cores = 4,
  replace = TRUE,
  seed = NULL
)
}
\arguments{
\item{gene_index}{Numeric vector representing the indices of the genes in
the network.}

\item{expression}{Numeric matrix representing gene expression, rows
represent genes, and columns represent samples.}

\item{phenotype1, phenotype2}{Numeric vectors representing the indices of
two phenotypes}

\item{bootstrap_iterations}{Number of bootstrap iterations to run for
computing the null distribution for calculating the p-value}

\item{parallel}{Boolean determining if calculation should be run in parallel,
on unix systems it uses the mclapply function, and on windows it creates
a socket cluster and uses parLapply.}

\item{cores}{number of cores to use for parallel computation if desired.
If it is greater than the number of cores available (as determined by
parallel::detectCores), it is instead set to the number of cores
available.}

\item{replace}{Whether the sampling should be done with replacement. If it is
TRUE, then two samples of size p1.size and p2.size respectively are taken
with replacement from combined (these two samples can overlap). If FALSE,
then a sample of size p1.size is taken without replacement from combined,
and the remainder of combined is taken as the other sample.}

\item{seed}{Integer, used to set the seed for the random number generator
used for sampling.}
}
\value{
A named list, with value equal to the absolute difference in rank
conservation scores between the two phenotypes, and pval equal to the
p-value found with bootstrapping.
}
\description{
\code{dirac.compare_phenotype} computes the absolute difference between the rank
conservation indices for two phenotypes based on a single gene network, and
the associated p-value using a bootstrapping approach.
}
\examples{
# example code
}
