% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirac.R
\name{dirac.compare_phenotype.single}
\alias{dirac.compare_phenotype.single}
\title{Compare two phenotypes rankings using the conservation index}
\usage{
dirac.compare_phenotype.single(rank_matrix, phenotype1, phenotype2)
}
\arguments{
\item{rank_matrix}{Boolean matrix, columns represent the rank vector for
a given sample}

\item{phenotype1, phenotype2}{Numeric vectors, indices of the samples in a
particular phenotype}
}
\value{
Float, absolute difference in rank conservation index between the
two phenotypes for a given gene network (implied by rank_matrix)
}
\description{
\code{dirac.compare_phenotype.single}computes the absolute difference between the
rank conservation index for a gene network between two phenotypes.
}
\examples{
# example code
}
