% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirac.R
\name{dirac.rank_conservation_index}
\alias{dirac.rank_conservation_index}
\title{Find rank conservation index}
\usage{
dirac.rank_conservation_index(rank_matrix, rank_template)
}
\arguments{
\item{rank_matrix}{Boolean matrix, columns represent each samples rank
vector}

\item{rank_template}{Boolean vector, represents template to match against}
}
\value{
Float representing the rank conservation index
}
\description{
\code{dirac.rank_conservation_index} finds the rank rank conservation index in
gene expression data
}
\examples{
expression <-  matrix(1:16, ncol=4, nrow=4)
rank_matrix <-dirac.rank_matrix(expression)
rank_template <- dirac.rank_template(rank_matrix)
dirac.rank_conservation_index(rank_matrix, rank_template)
}
