% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirac.R
\name{dirac.rank_difference_score}
\alias{dirac.rank_difference_score}
\title{Find the Rank Difference Score}
\usage{
dirac.rank_difference_score(expression, phenotype1, phenotype2, gene_index)
}
\arguments{
\item{expression}{Numeric matrix with gene expression data, rows are
genes, columns are samples.}

\item{phenotype1, phenotype2}{Numeric vectors containing the indexes for the
phenotypes.}

\item{gene_index}{Numeric vector of indices for the genes in the network}
}
\value{
A list of two vectors, each containing the rank difference scores
for the samples within the phenotypes.
}
\description{
\code{dirac.rank_difference_score} finds the rank difference score for a network
between 2 phenotypes. The score against the rank template for the second
phenotype is subtracted from the score against the rank template for the
first phenotype.
}
\examples{
expression=matrix(1:16, ncol=4, nrow=4)
dirac.rank_difference_score(expression, phenotype1=c(1,4), phenotype2=
 c(2,3), gene_index=c(1,2,4))
}
