% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirac.R
\name{dirac.rank_matching_score.vector}
\alias{dirac.rank_matching_score.vector}
\title{Find rank matching score across samples}
\usage{
dirac.rank_matching_score.vector(rank_matrix, rank_template)
}
\arguments{
\item{rank_matrix}{Boolean matrix, columns represent each samples rank
vector}

\item{rank_template}{Boolean vector, represents template to match against}
}
\value{
Numeric vector of rank matching scores
}
\description{
\code{dirac.rank_matching_score.vector} finds the rank matching score between each
sample in a rank_matrix, and a rank_template
}
\examples{
expression <-  matrix(1:16, ncol=4, nrow=4)
rank_matrix <-dirac.rank_matrix(expression)
rank_template <- dirac.rank_template(rank_matrix)
dirac.rank_matching_score.vector(rank_matrix, rank_template)
}
