% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirac.R
\name{dirac.rank_matrix}
\alias{dirac.rank_matrix}
\title{Determine DIRAC ranking matrix}
\usage{
dirac.rank_matrix(expression)
}
\arguments{
\item{expression}{A gene expression matrix, rows are genes, columns are
samples}
}
\value{
matrix whose columns represent the rank_vector for each sample
}
\description{
\code{dirac.rank_matrix} creates a DIRAC ranking matrix from expression data.
}
\examples{
dirac.rank_matrix(matrix(1:16, ncol=4, nrow=4))

}
