% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirac.R
\name{dirac.rank_template}
\alias{dirac.rank_template}
\title{Generate rank template}
\usage{
dirac.rank_template(rank_matrix)
}
\arguments{
\item{rank_matrix}{Rank matrix, from dirac.rank_matrix}
}
\value{
Numeric vector, rank template
}
\description{
\code{dirac.rank_template} creates a DIRAC ranking template from expression data.
}
\details{
Strict inequality is used for finding the rank template.
}
\examples{
dirac.rank_template(matrix(1:16,ncol=4, nrow=4))

}
