% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirac.R
\name{diracBootstrapScore}
\alias{diracBootstrapScore}
\title{Find a p-value for the difference in entropy between two phenotypes according
to the DIRAC method using a bootstrapped distribution}
\usage{
diracBootstrapScore(
  expression,
  geneNetwork,
  phenotype1,
  phenotype2,
  bootstrapIterations = 1000,
  replace = TRUE,
  asFrame = TRUE,
  BPPARAM = bpparam()
)
}
\arguments{
\item{expression}{Numeric matrix of gene expression, with rows representing
genes, and columns representing samples.}

\item{geneNetwork}{Index Vector, representing the indices of the genes in the
gene network}

\item{phenotype1, phenotype2}{Index vectors, representing the indices of the
phenotypes in the expression matrix}

\item{bootstrapIterations}{Integer, number of iterations to perform when
creating the null distribution}

\item{replace}{Boolean, whether sampling the phenotypes should be done with
replacement}
}
\value{
Named list of results, specifically p1Score, p2Score,
absoluteDifference, and pValue
}
\description{
Find a p-value for the difference in entropy between two phenotypes according
to the DIRAC method using a bootstrapped distribution
}
