% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirac.R
\name{diracClassificationRateBest}
\alias{diracClassificationRateBest}
\title{Compare classification rate for different gene networks}
\usage{
diracClassificationRateBest(expression, phenotype1, phenotype2, gene_index)
}
\arguments{
\item{expression}{Numeric matrix, represents gene expression, rows are genes
columns are samples}

\item{phenotype1, phenotype2}{Numeric vectors, contain the indices for the
samples corresponding to each phenotype.}

\item{gene_index}{List of numeric vectors, each containing the indices for
one of the gene networks.}

\item{parallel}{Boolean indicating Whether the computation should be run in
parallel. On unix makes use of the R core library parallel, on windows
requires foreach, and doParallel.}

\item{cores}{Integer, indicates number of cores to use for parallel
calculations. If greater than parallel::detectCores() return value,
will instead be set to said return value.}
}
\value{
String, name of best gene network
}
\description{
\code{dirac.classification_rate.best} finds the network with the best
classification rate between multiple gene networks for two phenotypes.
}
\examples{
# example code
}
