% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirac.R
\name{diracClassificationRateCompare}
\alias{diracClassificationRateCompare}
\title{Compare classification rate for different gene networks}
\usage{
diracClassificationRateCompare(
  expression,
  phenotype1,
  phenotype2,
  geneIndexList,
  asFrame = FALSE
)
}
\arguments{
\item{expression}{Numeric matrix, represents gene expression, rows are genes
columns are samples}

\item{phenotype1, phenotype2}{Numeric vectors, contain the indices for the
samples corresponding to each phenotype.}

\item{gene_index}{List of numeric vectors, each containing the indices for
one of the gene networks.}

\item{parallel}{Boolean indicating Whether the computation should be run in
parallel. On unix makes use of the R core library parallel, on windows
requires foreach, and doParallel.}

\item{cores}{Integer, indicates number of cores to use for parallel
calculations. If greater than parallel::detectCores() return value,
will instead be set to said return value.}

\item{as.frame}{Boolean, whether return value should be data.frame or named
list, see return for more information.}
}
\value{
if as.frame is TRUE:
data.frame with gene_network names as one column, and classification
rate as another.
if as.frame is FALSE:
Numeric vector containing the classification rate for each of the
provided gene networks. Named based on names of gene_index argument.
}
\description{
\code{dirac.classification_rate.compare} compares the classification rate between
multiple gene networks for two phenotypes
}
\examples{
# example code
}
