% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirac.R
\name{diracComparePhenotypes}
\alias{diracComparePhenotypes}
\title{Compare the entropy difference between phenotypes for a list of gene
networks using the DIRAC method}
\usage{
diracComparePhenotypes(
  expression,
  geneNetworkList,
  phenotype1,
  phenotype2,
  bootstrapIterations = 1000,
  replace = TRUE,
  asFrame = TRUE,
  BPPARAM = bpparam()
)
}
\arguments{
\item{expression}{Numeric matrix of gene expression values, rows
representing genes, and columns representing samples}

\item{geneNetworkList}{List of index vectors, each vector representing the
indices of a different gene networks}

\item{phenotype1, phenotype2}{Index vectors, representing the indices of the
phenotypes in the expression matrix}

\item{bootstrapIterations}{Integer, number of iterations to perform when
creating the null distribution}

\item{replace}{Boolean, whether sampling the phenotypes should be done with
replacement}

\item{asFrame}{Boolean, whether the return should be a data.frame (TRUE),
or a named list (FALSE)}
}
\value{
Either a data.frame, or a named list depending on the asFrame
parameter. In the data.frame form, with columns for geneNetwork, p1Score,
p2Score, absoluteDifference, and pValue. The list return is named
according to the names in the geneNetworkList parameter, with each value
being a named list with p1Score, p2Score, absoluteDifference, and pValue.
}
\description{
Compare the entropy difference between phenotypes for a list of gene
networks using the DIRAC method
}
