% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirac.R
\name{diracRankDifferenceScore}
\alias{diracRankDifferenceScore}
\title{Find the Rank Difference Score}
\usage{
diracRankDifferenceScore(expression, phenotype1, phenotype2, geneIndex)
}
\arguments{
\item{expression}{Numeric matrix with gene expression data, rows are
genes, columns are samples.}

\item{phenotype1, phenotype2}{Numeric vectors containing the indexes for the
phenotypes.}

\item{geneInex}{Numeric vector of indices for the genes in the network}
}
\value{
A list of two vectors, each containing the rank difference scores
for the samples within the phenotypes.
}
\description{
Finds the rank difference score for a network
between 2 phenotypes. The score against the rank template for the second
phenotype is subtracted from the score against the rank template for the
first phenotype.
}
\examples{
expression <- matrix(1:16, ncol = 4, nrow = 4)
dirac.rank_difference_score(expression,
    phenotype1 = c(1, 4), phenotype2 =
        c(2, 3), gene_index = c(1, 2, 4)
)
}
