% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirac.R
\name{diracRankFunction}
\alias{diracRankFunction}
\title{Determine DIRAC ranking matrix}
\usage{
diracRankFunction(filteredExpression)
}
\arguments{
\item{expression}{A gene expression matrix, rows are genes, columns are
samples}
}
\value{
matrix whose columns represent the rank_vector for each sample
}
\description{
Creates a DIRAC ranking matrix from expression data.
}
\examples{
dirac.rank_matrix(matrix(1:16, ncol = 4, nrow = 4))

}
