% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirac.R
\name{diracRankMatchingScore}
\alias{diracRankMatchingScore}
\title{Find rank matching score}
\usage{
diracRankMatchingScore(rankVector, rankTemplate)
}
\arguments{
\item{rank_vector}{Boolean vector representing the rank vector, must be
the same length as rank_template.}

\item{rank_template}{Boolean vector representing the rank template.}
}
\value{
Float, representing the rank matching score
}
\description{
Finds the rank matching score between a rank vector, and a rank template.
}
\examples{
dirac.rank_matching_score(c(1, 0, 0, 1, 0), c(0, 1, 0, 1, 0))
}
