% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirac.R
\name{diracRankVector}
\alias{diracRankVector}
\title{Determine DIRAC ranking vector}
\usage{
diracRankVector(expressionVector, expressionLength = NULL, matrixIndex = NULL)
}
\arguments{
\item{expression}{A numeric vector of gene expression data}

\item{expression.length}{Optional, length of expression matrix, useful for
repeated calls to this function with same expression length}

\item{matrix.index}{Optional, a Boolean vector or matrix with TRUE for
entries of gene-by-gene square matrix to be included in the rank_vector.
Mainly used for repeated calls to this function so it doesn't have to be
recalculated each time.}
}
\value{
Numeric rank vector
}
\description{
\code{dirac.rank_vector} creates a DIRAC ranking vector from expression data.
}
\examples{
dirac.rank_vector(c(4, 2, 1, 3))
dirac.rank_vector(c(4, 2, 1, 3), 4)
}
