% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirac.R
\name{diracSampleScore}
\alias{diracSampleScore}
\title{Compute Sample-wise entropy using DIRAC}
\usage{
diracSampleScore(filteredExpression)
}
\arguments{
\item{filteredExpression}{Numeric matrix, representing gene expression,
with rows representing genes in the network of interest, and columns
representing samples within a phenotype}
}
\value{
Numeric vector, DIRAC Rank Matching Score for each sample
}
\description{
Compute Sample-wise entropy using DIRAC
}
