% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirac.R
\name{diracScoreFunction}
\alias{diracScoreFunction}
\title{Score using DIRAC}
\usage{
diracScoreFunction(rankMatrix)
}
\arguments{
\item{rank_matrix}{Boolean matrix, columns represent each samples rank
vector}

\item{rank_template}{Boolean vector, represents template to match against}
}
\value{
Float representing the rank conservation index
}
\description{
Finds the rank rank conservation index in gene expression data
}
\examples{
expression <- matrix(1:16, ncol = 4, nrow = 4)
rankMatrix <- diracRankMatrix(expression)
rankTemplate <- diracRankTemplate(rankMatrix)
diracScoreFunction(rankMatrix, rankTemplate)
}
