% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{.craneEntropyClassifier}
\alias{.craneEntropyClassifier}
\title{Create a phenotype classifier based on CRANE entropy}
\usage{
.craneEntropyClassifier(expression, phenotype1, phenotype2, geneIndex)
}
\arguments{
\item{expression}{Numeric matrix of gene expression values, rows are genes
columns are samples}

\item{phenotype1, phenotype2}{Numeric vectors with indices for each of the
phenotypes}

\item{geneIndex}{Numeric vector of indices for the genes within the gene
network being used for classification}
}
\value{
A function with expression argument, which will return a logical
vector, with TRUE for the samples predicted to be phenotype1 and FALSE
for the samples predicted to be phenotype2
}
\description{
When the distance to the two rank centroids is equal, this will return
TRUE, so the ordering of phenotypes will affect classification
}
\examples{
# example code
}
