% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ensure_named}
\alias{ensure_named}
\title{Ensures that the provided list has names}
\usage{
ensure_named(in_list, prefix = "gn_")
}
\arguments{
\item{prefix}{string, Prefix for added names}

\item{l}{list, Input list to which will be returned with names, if it is not
already named,}
}
\value{
A named list, with the same data as the input but which is named
}
\description{
Ensures that the provided list has names
}
