% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen.R
\name{generateSummarizedExperiment}
\alias{generateSummarizedExperiment}
\title{Generates a summarized experiment object with gene expression data, which
has an ordered subnetwork}
\usage{
generateSummarizedExperiment(
  ngenesOrdered,
  ngenesTotal,
  nsamplesOrdered,
  nsamplesTotal,
  dist,
  reorderGenes = FALSE,
  genePrefix = "g_",
  samplePrefix = "s_",
  ...
)
}
\arguments{
\item{ngenesOrdered}{Integer, number of genes in the ordered sub network.}

\item{ngenesTotal}{Integer, total number of genes desired.}

\item{nsamplesOrdered}{Integer, number of samples in the ordered phenotype.}

\item{nsamplesTotal}{Integer, total number of samples}

\item{dist}{Function, distribution to use for generating data. It is passed
ngenes (for number of values to generate), and any other keyword
arguments. Must return a numeric vector.}

\item{reorderGenes}{Boolean, whether the order of genes within the ordered
sub network should be shuffled. If TRUE, the order of the genes in the
network will be randomly reshuffled. If FALSE, the genes within the network
will always have their expression values increase with increasing row
index.}

\item{...}{Keyword arguments, passed to the dist function}
}
\value{
Named list with:
seObject: The generated summarized experiment, with
colData:
sampleNames, phenotypeNum, and phenotypeStr
rowData:
geneNames, networkNum, netowrkStr
orderedGenes: A numeric vector with the indices of the ordered genes
orderedSamples: A numeric vector with the indices of the ordered samples
orderedGeneNames: A character vector with the ordered gene names
orderedSampleNames: A character vector with the ordered sample names
unorderedGenes: A numeric vector with the indices of the unordered genes
unorderedSamples: A numeric vector with the indices of the unorrdered
samples
unorderedGeneNames: A character vector with the unordered gene names
unorderedSampleNames: A character vector with the unordered sample
names
}
\description{
Generates a summarized experiment object with gene expression data, which
has an ordered subnetwork
}
