% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gnat-package.R
\docType{package}
\name{Gnat-package}
\alias{Gnat}
\alias{Gnat-package}
\title{Gnat: Analyze rank entropy in gene expression networks}
\description{
Functions for evaluating the rank entropy of gene networks based on gene expression data. Included methods based on DIRAC, and also the Kendall Rank Correlation Coefficient.
}
\author{
\strong{Maintainer}: Braden Griebel \email{bgriebel@uw.edu} (\href{https://orcid.org/0009-0003-3617-5734}{ORCID})

}
\keyword{internal}
