% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.R
\name{infer.compare_phenotypes}
\alias{infer.compare_phenotypes}
\title{Compute the rank entropy difference and p-value between two phenotypes for
a gene network}
\usage{
infer.compare_phenotypes(
  gene_network,
  expression,
  phenotype1,
  phenotype2,
  bootstrap_iterations = 1000,
  parallel = TRUE,
  cores = 4,
  replace = TRUE,
  seed = NULL
)
}
\arguments{
\item{gene_network}{Integer vector, the indices of the gene network within
the expression matrix}

\item{expression}{Numeric matrix, the gene expression data. Rows represent
genes, and columns represent samples.}

\item{phenotype1, phenotype2}{Integer vectors representing the indices of the
phenotypes withint the expression matrix}

\item{bootstrap_iterations}{Integer, number of bootstrap iterations to
perform}

\item{parallel}{Boolean, whether parallel operation is desired}

\item{cores}{Integer, number of cores to use for parallel operation}

\item{replace}{Boolean, whether to sample with replacement during
bootstrapping}

\item{seed}{Integer, seed to use for random number generation and sampling}
}
\value{
Named list with p1.rank_entropy, p2.rank_entropy,
absolute_difference, and p.value.
}
\description{
Compute the rank entropy difference and p-value between two phenotypes for
a gene network
}
