% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.R
\name{infer.compare_phenotypes.shuffle}
\alias{infer.compare_phenotypes.shuffle}
\title{Find the absolute difference between rank entropies for shuffled phenotypes}
\usage{
infer.compare_phenotypes.shuffle(
  i,
  rank_matrix,
  combined,
  p1.size,
  p2.size,
  replace = TRUE
)
}
\arguments{
\item{i}{Integer, Bootstrap iteration}

\item{rank_matrix}{Integer Matrix, represents the within sample ranks of gene
expression, rows representing genes, columns representing samples}

\item{combined}{Integer vector, represents the concatenation of the two
phenotype index vectors}

\item{p1.size, p2.size}{Integers, the number of samples within each phenotype}

\item{replace}{Boolean, whether sampling should be done with replacement.

If replace is TRUE, will take samples for each phenotype, so the two
phenotypes may overlap. If replace is FALSE, will take sample for phenotype1,
then use the remaining indices for phenotype2, so there won't be any overlap}
}
\value{
double, The absolute difference between the rank entropies of the
shuffled phenotypes
}
\description{
Find the absolute difference between rank entropies for shuffled phenotypes
}
