% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.R
\name{infer.compare_phenotypes.single}
\alias{infer.compare_phenotypes.single}
\title{Calculate the absolute difference in rank entropies between two phenotypes}
\usage{
infer.compare_phenotypes.single(rank_matrix, phenotype1, phenotype2)
}
\arguments{
\item{rank_matrix}{Integer matrix representing the gene expression ranks,
rows representing genes, and columns representing samples}

\item{phenotype1, phenotype2}{Integer vectors representing the indices of the
two phenotypes within the rank_matrix}
}
\value{
Float, the absolute value of the difference between the rank
entropies for each of the phenotypes.
}
\description{
\code{infer.compare_phenotypes.single} computes the rank entropy for each of two
phenotypes, and returns the absolute value of the difference between them
}
