% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.R
\name{infer.gene_network.entropy}
\alias{infer.gene_network.entropy}
\title{Find the entropy for gene network ranks}
\usage{
infer.gene_network.entropy(expression, gene_network, phenotype)
}
\arguments{
\item{expression}{Numeric matrix representing gene expression values,
with rows representing genes, and columns representing samples}

\item{gene_network}{Integer vector representing the indices of the genes
in the gene network within the expression matrix}

\item{phenotype}{Integer vector representing the indices of the samples
in the phenotype within the expression matrix}
}
\value{
Float, representing the rank entropy of the gene_network in the
given phenotype
}
\description{
\code{infer.gene_network.entropy} finds the rank entropy for a gene network
}
