% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.R
\name{infer.rank_matrix}
\alias{infer.rank_matrix}
\title{Compute the rank matrix}
\usage{
infer.rank_matrix(expression.filtered, margin = 2)
}
\arguments{
\item{expression.filtered}{Numeric matrix representing the gene expression,
with genes as rows and samples as columns. Genes should only be those
in the network of interest, and samples should only be those in phenotype
of interest. Can function with the transpose of this by changing
margin to 2, but other methods don't work with this alternate
orientation.}

\item{margin}{Integer representing whether function rank should be applied
along rows (1) or columns (2)}
}
\value{
Integer array representing the ranks of the gene expression within
a sample.
}
\description{
Compute the rank matrix
}
