% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.R
\name{INFER}
\alias{INFER}
\title{Compute the rank entropy and difference probability for a list of gene
networks}
\usage{
INFER(
  expression,
  gene_network_list,
  phenotype1,
  phenotype2,
  bootstrap_iterations = 1000,
  parallel = TRUE,
  cores = 4,
  replace = TRUE,
  seed = NULL,
  as.frame = TRUE
)
}
\arguments{
\item{expression}{Numeric matrix representing gene expression data,
rows represent genes and columns represent samples}

\item{gene_network_list}{List of integer vectors, each vector is the
indices for a genes network within the expression matrix. The names of
this list are used for naming the gene networks.}

\item{phenotype1, phenotype2}{Integer vectors representing the indices of the
two phenotypes within the expression matrix.}

\item{bootstrap_iterations}{Integer, number of bootstrap_iterations to
perform}

\item{parallel}{Boolean, whether the calculation should be performed in
parallel. See details for more information.}

\item{cores}{Integer, number of cores to use for parallel operations}

\item{replace}{Boolean, whether to sample with replacement during the
bootstrapping.}

\item{seed}{Integer, seed to use for the phenotype sampling.}

\item{as.frame}{Boolean, whether return should be a dataframe or not. If
TRUE, will return a dataframe with gene_network, p1.rank_entropy,
p2.rank_entropy, absolute_difference, and p.value as columns. If FALSE,
will return this as a named list instead.}
}
\value{
dataframe or named list depending on as.frame.
If as.frame is TRUE:
returns a dataframe with gene_network, p1.rank_entropy,
p2.rank_entropy, absolute_difference and p.value columns.
If as.frame is FALSE:
returns the data as a named list. Names are the same as names for
gene_network_list argument, and each element is a named list with
gene_network, p1.rank_entropy, p2.rank_entropy, absolute_difference, and
p.value.

For parallel operation, on unix the parallel mclapply is used, which employs
a fork cluster. On windows, a psock cluster is used.
}
\description{
Compute the rank entropy and difference probability for a list of gene
networks
}
