% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.R
\name{inferGeneEntropy}
\alias{inferGeneEntropy}
\title{Find the rank entropy for each gene in a network}
\usage{
inferGeneEntropy(filteredExpression)
}
\arguments{
\item{filteredExpression}{Numeric matrix representing gene expression, with
rows representing genes in the network and columns representing samples
in the phenotype.}
}
\value{
Numeric vector representing the entropy for each gene in the network
in the given phenotype.
}
\description{
\code{inferGeneEntropy} Determines the rank entropy for each gene in a network
}
