% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.R
\name{inferRankFunction}
\alias{inferRankFunction}
\title{Rank function for the INFER method, rank each value in a column}
\usage{
inferRankFunction(filteredExpression)
}
\arguments{
\item{filteredExpression}{Filtered expression, where each row is a gene
within the gene network, and each column is a sample}
}
\value{
Numeric matrix, with each value representing the rank of the
expression within a sample compared to the other genes in the network
}
\description{
Rank function for the INFER method, rank each value in a column
}
\examples{
testMat <- matrix(runif(20), ncol=4, nrow=5)
print(craneRankFunction(testMat))
print(craneRankFunction(testMat, margin=1))
}
