% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.R
\name{inferScoreFunction}
\alias{inferScoreFunction}
\title{Function to score a rank matrix using the INFER method, the average
information entropy of genes within the network}
\usage{
inferScoreFunction(rankMatrix)
}
\arguments{
\item{rankMatrix}{Numeric matrix, a matrix where each value represents
the rank of the gene's expression within the gene network}
}
\value{
Numeric score, value representing the mean information entropy
for the genes within the network in the given phenotype
}
\description{
Function to score a rank matrix using the INFER method, the average
information entropy of genes within the network
}
