% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/race.R
\name{race.compare_phenotypes.shuffle}
\alias{race.compare_phenotypes.shuffle}
\title{Compare shuffled Phenotypes}
\usage{
race.compare_phenotypes.shuffle(
  i,
  expression.gene_network,
  combined,
  p1.size,
  p2.size,
  replace = TRUE
)
}
\arguments{
\item{i}{Bootstrap iteration}

\item{expression.gene_network}{Numeric matrix, represents the gene
expression, with rows representing genes in the network of interest,
and columns representing samples.}

\item{combined}{Integer vector, represents the indices of both phenotypes
(concatenated) within the expression.gene_network matrix}

\item{p1.size, p2.size}{Integers representing the size of the two phenotypes}

\item{replace}{Boolean, whether sampling should be done with replacement}
}
\value{
Double, the absolute difference between the mean Kendall
Rank Correlation Coefficient of the two shuffled phenotypes
}
\description{
\code{race.compare_phenotypes.shuffle} Computes the absolute difference in
mean Kendall Rank Correlation Coefficient between shuffled phenotypes
}
