% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/race.R
\name{race.compare_phenotypes.single}
\alias{race.compare_phenotypes.single}
\title{Compare two phenotypes mean Kendall Correlation Coefficient}
\usage{
race.compare_phenotypes.single(expression.gene_network, phenotype1, phenotype2)
}
\arguments{
\item{expression.gene_network}{Numeric matrix, represents the gene
expression, with rows representing genes within the network of interest,
and columns representing the samples.}

\item{phenotype1, phenotype2}{Integer vectors, representing the indices of the
two phenotypes within the expression.gene_network matrix}
}
\value{
Double, the absolute difference between the mean Kendall Rank
Correlation of the two phenotypes
}
\description{
\code{race.compare_phenotypes.single} computes the absolute difference between the
mean Kendall rank correlation coefficient of the two phenotypes
}
