% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/race.R
\name{RACE}
\alias{RACE}
\title{Compare phenotypes and gene networks using RACE}
\usage{
RACE(
  expression,
  gene_network_list,
  phenotype1,
  phenotype2,
  bootstrap_iterations = 1000,
  parallel = TRUE,
  cores = 4,
  replace = TRUE,
  seed = NULL,
  as.frame = TRUE
)
}
\arguments{
\item{expression}{Numeric matrix, represents gene expression with rows
representing genes, and columns representing samples}

\item{gene_network_list}{Named list of integer vectors, each representing
the indices of the genes in the network within the expression matrix.}

\item{phenotype1, phenotype2}{Integer vectors representing the indices
of the samples in the phenotype within the expression matrix}

\item{bootstrap_iterations}{Integer number of iterations for the
bootstrapping}

\item{parallel}{Boolean, whether computations should be done in parallel.
See details for more information.}

\item{cores}{Integer, number of cores to use for parallel operation.}

\item{replace}{Boolean, whether sampling should be done with replacement}

\item{seed}{Integer, seed for RNG (note, even for the same seed, the
RNG is handled differently for parallel operation between different
operating systems).}

\item{as.frame}{Boolean, whether the result should be returned as a
dataframe or a named list

For parallel operations, on unix operating systems a fork cluster will be
used, while on windows a psock cluster will be used. The RNG for both uses
the L'Ecuyer method. For serial operation, a Mersenne-Twister RNG is used.
This means that results will vary between serial and parallel operation.}
}
\value{
A named list or dataframe depending on \code{as.dataframe}.
If as.dataframe is TRUE:
Will return a dataframe with a row for each of the gene networks,
and columns for gene_network, p1.mean_rank_correlation,
p2.mean_rank_correlation, absolute_difference, and p.value
If as.dataframe is FALSE:
will return a named list, named according to the names of the
\code{gene_network_list} argument. Each element will also be a named list,
with p1.mean_rank_correlation, p2.mean_rank_correlation,
absolute_difference, and p.value.
}
\description{
Compare phenotypes and gene networks using RACE
}
