% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/race.R
\name{raceRankFunction}
\alias{raceRankFunction}
\title{Rank function for the RACE method, simply the identity function}
\usage{
raceRankFunction(filteredExpression)
}
\arguments{
\item{filteredExpression}{Filtered expression, where each row is a gene
within the gene network, and each column is a sample}
}
\value{
Numeric matrix, identical to input
}
\description{
Rank function for the RACE method, simply the identity function
}
\details{
This function just returns the same matrix that it is provided, it is used
for providing compatability with the other methods
}
\examples{
testMat <- matrix(runif(20), ncol=4, nrow=5)
print(raceRankFunction(testMat))
print(raceRankFunction(testMat, margin=1))
}
