% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/race.R
\name{raceSampleScore}
\alias{raceSampleScore}
\title{Calculate the by sample entropy scores using the RACE method}
\usage{
raceSampleScore(filteredExpression)
}
\arguments{
\item{filteredExpression}{Numeric Matrix, rows representing genes in
the gene network, columns representing samples in the phenotype}
}
\value{
Numeric vector, with a score for the entropy of each sample
}
\description{
Calculate the by sample entropy scores using the RACE method
}
