% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/race.R
\name{raceScoreFunction}
\alias{raceScoreFunction}
\title{Function to score a rank matrix using RACE method, the average of the
Kendall Tau rank correlation coefficient between every pair of samples
in a phenotype}
\usage{
raceScoreFunction(rankMatrix)
}
\arguments{
\item{rankMatrix}{Numeric matrix, for this method this is the same as the
expression matrix}
}
\value{
Numeric score, value representing the mean Kendall Tau between all
pairs of samples within a phenotype
}
\description{
Function to score a rank matrix using RACE method, the average of the
Kendall Tau rank correlation coefficient between every pair of samples
in a phenotype
}
