% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rce.R
\name{rce.compare_phenotypes}
\alias{rce.compare_phenotypes}
\title{Compare phenotypes using Rank Correlation Entropy}
\usage{
rce.compare_phenotypes(
  gene_index,
  gene_expression,
  phenotype1,
  phenotype2,
  summary.function = stats::median,
  summary.name = "median",
  phenotype.names = NULL,
  bootstrap_iterations = 1000,
  parallel = TRUE,
  cores = 4,
  replace = TRUE,
  seed1 = NULL,
  seed2 = NULL
)
}
\arguments{
\item{gene_index}{Integer vector, represents the indices of genes within
a gene network.}

\item{gene_expression}{Numeric matrix representing gene expression, rows
represent genes, and columns represent samples.}

\item{phenotype1, phenotype2}{Integer vectors, representing the indices
of the two phenotypes}

\item{summary.function}{Function, takes a vector of floats and returns a
float. Will be passed the distribution generated by bootstrapping and
result will}

\item{bootstrap_iterations}{Integer, number of iterations to perform for
bootstrapping to create the distribution.}

\item{parallel}{Boolean, whether to use parallel computing. On unix systems
will use a fork cluster, on windows will use a psock cluster.}

\item{cores}{Integer, number of cores to use for parallel computation.}

\item{replace}{Boolean, whether to sample with replacement when performing
bootstrapping.}

\item{seed1}{Seed for bootstrapping the phenotype1 distribution.}

\item{seed2}{Seed for bootstrapping the phenotype2 distribution.}
}
\value{
List of the p-value, and the summary statistic for each of the
phenotype distributions. The p-value comes fom the two sample
Kolmogorov-Smirnov test, where the null and alternative hypotheses are:
Null hypothesis: The two distributions of rank correlations are from the
same distributions.
Alternate hypothesis: The two distributions of rank correlations are from
different distributions.
}
\description{
Compare phenotypes using Rank Correlation Entropy
}
\examples{
# example code

}
