% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rce.R
\name{rce.distribution}
\alias{rce.distribution}
\title{Compute empirical distribution for the Kendall Correlation Coefficient}
\usage{
rce.distribution(
  gene_expression,
  gene_index,
  phenotype,
  iterations = 1000,
  parallel = TRUE,
  cores = 4,
  replace = TRUE,
  seed
)
}
\arguments{
\item{gene_expression}{Numeric matrix, rows represent genes, and columns
represent samples}

\item{gene_index}{Integer vector, representing the indices in the gene
network}

\item{phenotype}{Integer vector, representing the indices of the phenotype}

\item{parallel}{Boolean, weather parallel operation is desired}

\item{cores}{Integer, number of  cores to use for parallel operation,
if greater than the number of available cores will be reduced to the
number of available cores}

\item{replace}{Boolean, Whether replacement should be used when sampling
phenotype for templates.}

\item{seed}{Integer, seed for random number generator}
}
\description{
\code{rce.distribution} calculates the empirical distribution of the Kendall
Correlation Coefficient.
}
\details{
For each iteration a template sample is chosen, and the Kendall rank
correlation coefficient is calculated between that template and all other
samples. Then the mean of these values is taken.
}
