% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rce.R
\name{rce.rank_correlation.matrix}
\alias{rce.rank_correlation.matrix}
\title{Compute the rank correlation entropy for one template}
\usage{
rce.rank_correlation.matrix(gene_expression, template)
}
\arguments{
\item{gene_expression}{Numeric matrix, rows are genes in the network,
and columns are samples}

\item{template}{Numeric vector,}
}
\value{
Float, mean of the Kendall Tau Rank correlation coefficients
between the template and all columns of gene_expression
}
\description{
\code{rce.rank_correlation.matrix} computes the Kendall Tau Rank correlation
coefficient between each column of gene_expression and the template
}
