% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rce.R
\name{rce.rank_correlation.vector}
\alias{rce.rank_correlation.vector}
\title{Compute the rank correlation entropy between two vectors}
\usage{
rce.rank_correlation.vector(x, y)
}
\arguments{
\item{x, y}{Numeric vectors between which the correlation coefficient is to be
calculated.}
}
\value{
Float, the Kendall Rank Correlation Coefficient between \code{x} and \code{y}
}
\description{
\code{rce.rank_correlation.vector} computes the Kendall Tau Rank Correlation
coefficient between two vectors
}
