% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rce.R
\name{rce}
\alias{rce}
\title{Calculate p-values for Rank Conservation Entropy}
\usage{
rce(
  gene_expression,
  gene_index_list,
  phenotype1,
  phenotype2,
  iterations = 1000,
  parallel = TRUE,
  cores = 4,
  replace = TRUE,
  seed1 = NULL,
  seed2 = NULL,
  as.frame = TRUE
)
}
\arguments{
\item{gene_expression}{Numeric matrix representing gene expression, rows are
genes, and columns are samples}

\item{gene_index_list}{List of Integer vectors, each representing the indices
of a gene network in the gene_expression matrix}

\item{phenotype1, phenotype2}{Integer vectors representing the indices of the
two phenotypes in the gene_expression matrix}

\item{iterations}{Integer, Number of templates to sample from each of the
phenotypes to calculate the distribution of the Kendall Rank Correlation
coefficient}

\item{parallel}{Boolean describing whether parallel operation is desired. On
unix systems creates fork cluster, on windows creates a psock cluster}

\item{cores}{Integer describing number of cores to use for parallel
operation.}

\item{replace}{Boolean, describing whether the phenotype samples should be
taken with replacement}

\item{seed1, seed2}{Seeds to use for sampling the respective phenotypes}

\item{as.frame}{Boolean, describing whether the results should be returned
as a data.frame}
}
\value{
If as.frame is TRUE:
returns a dataframe with columns of gene network names
(\verb{$network_names}), and p-values (\verb{$pvals})
If as.frame is FALSE:
returns a named list of p-values, with names corresponding to names
of the gene_index_list
}
\description{
For each of the gene networks described by the gene_index_list, calculates
the p-values for the difference in rank conservation entropy between
the two phenotypes.
}
\examples{
# Create example random expression data
expression <- matrix(runif(50*120, min=1, max=500000), ncol=50)
# Create list of integer vectors representing different gene networks
gene_index_list <- list(
  gene_network_A = c(34,35,36,40,47,101),
  gene_network_B = c(115,116,120),
  gene_network_C = c(1,5,17,19,20,25,31,33)
)
# Create integer vector representing indices of phenotype1 and phenotype2
phenotype1 <- sample(50, 20,replace=FALSE)
phenotype2 <- sample(50, 15,replace=FALSE)
# Run the RCE
results <- rce(expression, gene_index_list, phenotype1, phenotype2,
  iterations=100, parallel=FALSE, replace=TRUE, seed1=42, seed2=314,
  as.frame=TRUE)
# Show results
print(results)
# Repeat using parallel operation
results <- rce(expression, gene_index_list, phenotype1, phenotype2,
  iterations=100, parallel=FALSE, cores=2, replace=TRUE, seed1=42, seed2=314,
  as.frame=TRUE)
# View the results, note likely not the same due to small number of
# iterationsand the different ways that random number generation is handled
# between serial and parallel operation
print(results)
}
