% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rce.R
\name{rce.sample}
\alias{rce.sample}
\title{Compute the Kendall correlation coefficient between gene expression and
sampled template}
\usage{
rce.sample(template_index, gene_expression)
}
\arguments{
\item{template_index}{Integer, index of sample to use as template}

\item{gene_expression}{Numeric matrix representing the gene expression, with
genes in the network as rows, and samples as columns}
}
\value{
Float, the mean of the Kendall Rank correlation coefficient
between the sampled template and the other samples in the phenotype
}
\description{
\code{rce.sample} Takes in gene expression matrix, and index for phenotype to be
used as template, then computes the Kendall Rank Correlation Coefficient
between all other samples and the template
}
