% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_score.R
\name{scoreShuffled}
\alias{scoreShuffled}
\title{Compare two shuffled phenotypes}
\usage{
scoreShuffled(
  iteration,
  rankMatrix,
  combined,
  p1Size,
  p2Size,
  scoreFun,
  replace = TRUE
)
}
\arguments{
\item{iteration}{Value representing the current bootstrap iteration,
used to make signature match the required signature for \code{bplapply}.}

\item{rankMatrix}{Numeric matrix with samples as the columns. See details
for further information.}

\item{combined}{Vector representing the concatenated indices of
phenotype 1 and 2.}

\item{p1Size, p2Size}{Integer representing the size of the two phenotypes}

\item{scoreFun}{Function which calculates a score form the
rank matrix.}

\item{replace}{Boolean representing Whether sampling should be done with
replacement}
}
\description{
Compare two shuffled phenotypes
}
\details{
Rank matrix is normally generated from a passed rank_fun function, so the
columns should represent samples, but what the rows represent depends
on the specific rank function. For RACE, CRANE, and INFER the rows simply
represent genes, while for DIRAC they represent specific pairwise
comparison.

The combined vector can contain any values that can be used with the
\code{[} matrix function, as well as the vector \code{c} and \code{sample} functions.

The replace parameter dictates whether the sampling to create shuffled
phenotype index vectors should be done with replacement. If TRUE, sampling
with replacement if performed to create the two shuffled index vectors
of the appropriate sizes. If FALSE, random sampling without replacement
occurs once to select phenotype1 indices, and the remaining indices are used
for phenotype2.
}
