% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seIntegration.R
\name{seCrane}
\alias{seCrane}
\title{Perform Phenotype Comparison using the CRANE method on a Summarized
Experiment Object}
\usage{
seCrane(
  seObject,
  phenotype1,
  phenotype2,
  geneNetworkList,
  assayName = "counts",
  bootstrapIterations = 1000,
  replace = TRUE,
  asFrame = TRUE,
  BPPARAM = bpparam()
)
}
\arguments{
\item{seObject}{SummarizedExperiment object used as input, should contain
an assay with gene expression data, and genes as rownames}

\item{phenotype1, phenotype2}{Index vector describing the location of the
two phenotypes in the seObject, can be a numeric, character, or logical
vector}

\item{geneNetworkList}{(Named) list of gene network index vectors, each
describing the indices of a gene network, these can be integer,
character, or logical vectors}

\item{assayName}{Name of the assay containing the gene expression data}

\item{bootstrapIterations}{Number of iterations to use for bootstrapping
the null distribution}

\item{replace}{Logical determining if the permutation sampling done as
part of the bootstrapping should be done with or without replacement
(TRUE for replacement, FALSE for no replacement)}

\item{asFrame}{Logical determining if the return value should be in the form
of a data.frame}
}
\value{
Either a data.frame, or a named list depending on the asFrame
parameter. The data.frame form will have columns for geneNetwork, p1Score,
p2Score, absoluteDifference, and pValue. The list return is named
according to the names in the geneNetworkList parameter, with each value
being a named list with p1Score, p2Score, absoluteDifference, and pValue.
}
\description{
Perform Phenotype Comparison using the CRANE method on a Summarized
Experiment Object
}
