% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seIntegration.R
\name{seRaceBySample}
\alias{seRaceBySample}
\title{Compute the Sample Wise entropy using RACE for a SummarizedExperiment}
\usage{
seRaceBySample(
  seObject,
  phenotype1,
  phenotype2,
  phenotypeNames = c("phenotype1", "phenotype2"),
  geneNetwork,
  geneNetworkName = "gn",
  assayName = "counts"
)
}
\arguments{
\item{seObject}{SummarizedExperiment object used as input, should contain
an assay with gene expression data, and genes as rownames}

\item{phenotype1, phenotype2}{Index vector describing the location of the
two phenotypes in the seObject, can be a numeric, character, or logical
vector}

\item{phenotypeNames}{Character vector with names for the phenotypes,
phenotype1 should be the first entry and phenotype2 should be the
second}

\item{geneNetwork}{Index vector describing the location of the
gene network in the seObject, can be a numeric, character, or logical
vector}

\item{geneNetworkName}{Name for the gene network}

\item{assayName}{Name of the assay containing the gene expression data}
}
\value{
SummarizedExperiment object with two columns added to colData,
one for the entropy of each of the phenotypes
}
\description{
Compute the Sample Wise entropy using RACE for a SummarizedExperiment
}
