% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simpleRank}
\alias{simpleRank}
\title{Return a matrix where each value has been replaces by that values rank}
\usage{
simpleRank(filteredExpression, margin = 2)
}
\arguments{
\item{filteredExpression}{Numeric matrix to convert to a rank matrix}

\item{margin}{Integer, margin to rank along, 1 for rows, 2 for columns}
}
\value{
Numeric matrix of ranks
}
\description{
Return a matrix where each value has been replaces by that values rank
}
\examples{
testMat <- matrix(runif(20), ncol=4, nrow=5)
print(simpleRank(testMat))
print(simpleRank(testMat, margin=1))
}
