% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_score.R, R/compare_phenotypes.R
\name{comparePhenotypes}
\alias{comparePhenotypes}
\title{Title}
\usage{
comparePhenotypes(
  expression,
  method = c("CRANE", "DIRAC", "INFER", "RACE"),
  phenotype1,
  phenotype2,
  geneNetworkList,
  assayName = "counts",
  bootstrapIterations = 1000,
  replace = TRUE,
  asFrame = TRUE,
  BPPARAM = bpparam()
)

comparePhenotypes(
  expression,
  method = c("CRANE", "DIRAC", "INFER", "RACE"),
  phenotype1,
  phenotype2,
  geneNetworkList,
  assayName = "counts",
  bootstrapIterations = 1000,
  replace = TRUE,
  asFrame = TRUE,
  BPPARAM = bpparam()
)
}
\arguments{
\item{expression}{Numeric matrix of gene expression values, rows
representing genes, and columns representing samples}

\item{phenotype1, phenotype2}{Index vectors, representing the indices of the
phenotypes in the expression matrix}

\item{geneNetworkList}{List of index vectors, each vector representing the
indices of a different gene networks}

\item{bootstrapIterations}{Integer, number of iterations to perform when
creating the null distribution}

\item{replace}{Boolean, whether sampling the phenotypes should be done with
replacement}

\item{asFrame}{Boolean, whether the return should be a data.frame (TRUE),
or a named list (FALSE)}

\item{BPPARAM}{}

\item{rankFun}{Function for creating a rank matrix, should take a matrix of
dim (gene, sample) as input, and return a rank matrix, with samples as the
columns.}

\item{scoreFun}{Function for scoring entropy in a phenotype, should take a
rank matrix as input, and return a score (the rank matrix will only
include samples within 1 phenotype)}
}
\value{
Either a data.frame, or a named list depending on the asFrame
parameter. In the data.frame form, with columns for geneNetwork, p1Score,
p2Score, absoluteDifference, and pValue. The list return is named
according to the names in the geneNetworkList parameter, with each value
being a named list with p1Score, p2Score, absoluteDifference, and pValue.
}
\description{
Title
}
