% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirac.R
\name{dirac.classification_rate}
\alias{dirac.classification_rate}
\title{Find the classification rate for single network}
\usage{
dirac.classification_rate(expression, phenotype1, phenotype2, gene_index)
}
\arguments{
\item{expression}{Numeric matrix with gene expression data, rows are genes
columns are samples.}

\item{phenotype1, phenotype2}{Numeric vectors containing the indexes for the
two phenotypes to compare.}

\item{gene_index}{Numeric vector of index for genes in network}
}
\value{
List of floats, representing the classification rate for each
network. Names will be taken from the gene_index list.
}
\description{
\code{dirac.classification_rate} Finds the classification rate between two
phenotypes based on a gene network.
}
\details{
Classification rate is defined as the average between sensitivity and
specificity. classification_rate =
Pr(rank_diff>0|phenotype=A)*Pr(phenotype=A) +
Pr(rank_diff<=0|phenotype=B)*Pr(phenotype=B)
}
\examples{
# example code
}
