% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{diracClassifier}
\alias{diracClassifier}
\title{Creates a phenotype classifier based on DIRAC}
\usage{
diracClassifier(expression, phenotype1, phenotype2, geneIndex)
}
\arguments{
\item{expression}{Numeric matrix of gene expression values, rows are genes
columns are samples}

\item{phenotype1, phenotype2}{Numeric vectors with indices for each of the
phenotypes}

\item{gene_index}{Numeric vector of indices for the genes within the gene
network being used for classification}

\item{names}{optional Character vector with the desired names for the
phenotypes}
}
\value{
A function with expression argument, which will return a boolean
vector, with TRUE for the samples predicted to be phenotype1 and FALSE
for the samples predicted to be phenotype2
}
\description{
Creates a phenotype classifier based on DIRAC
}
\examples{
# example code
}
