% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gnat-package.R
\docType{package}
\name{Gnat-package}
\alias{Gnat}
\alias{Gnat-package}
\title{Gnat: Analyze rank entropy in gene expression networks}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Braden Griebel \email{bgriebel@uw.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
