% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rce.R
\name{rce.compare_phenotypes}
\alias{rce.compare_phenotypes}
\title{Compare phenotypes using Rank Correlation Entropy}
\usage{
rce.compare_phenotypes(
  gene_index,
  gene_expression,
  phenotype1,
  phenotype2,
  bootstrap_iterations = 1000,
  parallel = TRUE,
  cores = 4,
  replace = TRUE,
  seed1,
  seed2
)
}
\arguments{
\item{gene_index}{Integer vector, represents the indices of genes within
a gene network.}

\item{gene_expression}{Numeric matrix representing gene expression, rows
represent genes, and columns represent samples.}

\item{phenotype1, phenotype2}{Integer vectors, representing the indices
of the two phenotypes}

\item{bootstrap_iterations}{Integer, number of iterations to perform for
bootstrapping to create the distribution.}

\item{parallel}{Boolean, whether to use parallel computing. On unix systems
will use a fork cluster, on windows will use a psock cluster.}

\item{cores}{Integer, number of cores to use for parallel computation.}

\item{replace}{Boolean, whether to sample with replacement when performing
bootstrapping.}

\item{seed1}{Seed for bootstrapping the phenotype1 distribution.}

\item{seed2}{Seed for bootstrapping the phenotype2 distribution.}
}
\value{
Float representing the p-value,
Null hypothesis: The two distributions of rank correlations are from the
same distributions.
Alternate hypothesis: The two distributions of rank correlations are from
different distributions.
}
\description{
Compare phenotypes using Rank Correlation Entropy
}
\examples{
# example code

}
