% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rce.R
\name{rce}
\alias{rce}
\title{Calculate p-values for Rank Conservation Entropy}
\usage{
rce(
  gene_expression,
  gene_index_list,
  phenotype1,
  phenotype2,
  iterations = 1000,
  parallel = TRUE,
  cores = 4,
  replace = TRUE,
  seed1,
  seed2,
  as.frame = TRUE
)
}
\arguments{
\item{gene_expression}{Numeric matrix representing gene expression, rows are
genes, and columns are samples}

\item{gene_index_list}{List of Integer vectors, each representing the indices
of a gene network in the gene_expression matrix}

\item{phenotype1, phenotype2}{Integer vectors representing the indices of the
two phenotypes in the gene_expression matrix}

\item{iterations}{Integer, Number of templates to sample from each of the
phenotypes to calculate the distribution of the Kendall Rank Correlation
coefficient}

\item{parallel}{Boolean describing whether parallel operation is desired. On
unix systems creates fork cluster, on windows creates a psock cluster}

\item{cores}{Integer describing number of cores to use for parallel
operation.}

\item{replace}{Boolean, describing weather the phenotype samples should be
taken with replacement}

\item{seed1, seed2}{Seeds to use for sampling the respective phenotypes}

\item{as.frame}{Boolean, describing weather the results should be returned
as a data.frame}
}
\value{
If as.frame is TRUE:
returns a dataframe with columns of gene network names
(\verb{$network_names}), and p-values (\verb{$pvals})
If as.frame is FALSE:
returns a named list of p-values, with names corresponding to names
of the gene_index_list
}
\description{
For each of the gene networks described by the gene_index_list, calculates
the p-values for the difference in rank conservation entropy between
the two phenotypes.
}
